+++ sanctuary-theme/widgets/class-event-cards.php
@@ -27,7 +27,7 @@ class Sanctuary_Widget_EventCards extends Sanctuary_Base_Widget {
 		return 'eicon-posts-grid';
 	}
 
-	protected function register_controls() {
+	protected function register_content_controls() {
 		$this->start_controls_section( 'content', array( 'label' => __( 'Content', 'sanctuary' ) ) );
 
 		$this->add_heading_controls(