+++ sanctuary-theme/widgets/class-gallery.php
@@ -27,7 +27,7 @@ class Sanctuary_Widget_Gallery extends Sanctuary_Base_Widget {
 		return 'eicon-gallery-masonry';
 	}
 
-	protected function register_controls() {
+	protected function register_content_controls() {
 		$this->start_controls_section( 'content', array( 'label' => __( 'Content', 'sanctuary' ) ) );
 
 		$this->add_heading_controls(