+++ sanctuary-theme/widgets/class-included-list.php
@@ -27,7 +27,7 @@ class Sanctuary_Widget_IncludedList extends Sanctuary_Base_Widget {
 		return 'eicon-bullet-list';
 	}
 
-	protected function register_controls() {
+	protected function register_content_controls() {
 		$this->start_controls_section( 'content', array( 'label' => __( 'Content', 'sanctuary' ) ) );
 
 		$this->add_control( 'bg', array(