+++ sanctuary-theme/widgets/class-instagram.php
@@ -28,7 +28,7 @@ class Sanctuary_Widget_Instagram extends Sanctuary_Base_Widget {
 		return 'eicon-instagram-gallery';
 	}
 
-	protected function register_controls() {
+	protected function register_content_controls() {
 		$this->start_controls_section( 'content', array( 'label' => __( 'Content', 'sanctuary' ) ) );
 
 		$this->add_control( 'label', array(