+++ sanctuary-theme/widgets/class-rates.php
@@ -25,7 +25,7 @@ class Sanctuary_Widget_Rates extends Sanctuary_Base_Widget {
 		return 'eicon-price-tag';
 	}
 
-	protected function register_controls() {
+	protected function register_content_controls() {
 		$this->start_controls_section( 'content', array( 'label' => __( 'Content', 'sanctuary' ) ) );
 
 		$this->add_control( 'eyebrow', array( 'label' => __( 'Eyebrow', 'sanctuary' ), 'type' => Controls_Manager::TEXT, 'default' => __( 'Rates', 'sanctuary' ) ) );